changing data type of itemCount to long

+++ lib/jdomodels/src/main/java/org/sagebionetworks/repo/model/dbo/dao/NodeDAOImpl.java
@@ -1878,7 +1878,7 @@ public class NodeDAOImpl implements NodeDAO, InitializingBean {
 					}
 					Object count = entityPropertyAnnotation.getSingleValue("count");
 					if (count != null) {
-						dto.setItemCount(Integer.parseInt(count.toString()));
+						dto.setItemCount(Long.parseLong(count.toString()));
 					}
 				}
